+++ lib/src/models/device_online_state.dart
@@ -1,0 +1,7 @@
+part of '../../viam_flutter_provisioning_widget.dart';
+
+enum DeviceOnlineState {
+  checking,
+  agentConnected,
+  success,
+}
